+++ src/ulong_extras/test/t-powmod2_preinv.c
@@ -42,7 +42,7 @@ TEST_FUNCTION_START(n_powmod2_preinv, state)
         flint_mpz_set_ui(d_m, d);
         if (exp < 0)
         {
-            flint_mpz_powm_ui(r2_m, a_m, -exp, d_m);
+            flint_mpz_powm_ui(r2_m, a_m, -(ulong) exp, d_m);
             mpz_invert(r2_m, r2_m, d_m);
         }
         else
